feat: :sparkles: código adaptado para gerenciamento de estados
